Finished weird scatter matrix visual for nD cases.
I initially thought I was clever plotting a matrix of each dimension
against another- and it works nice for 2D data, but it doesn't really
make sense past this. It draws the hyperplane for each respective axis
and weight set. *shrug* Also included is a partial implementation of a
parallel coordinate graph.